cleaned up main directory, condensed plot scripts to a single script
