+++ README.md
@@ -1,15 +1,15 @@
 # SPC-Conv-Forecast
 ## Example
-Sample Day 1 SPC Forecast: (05/18/2017 1200Z Day 1 Forecast) ![picture alt](https://raw.githubusercontent.com/tmaldrsn/SPC-Conv-Forecast/v0.3.4/2017-05-18_1200_UTC_Day_1_convective_outlook%20(1).gif "SPC Day 1 05/18/2017 1200Z Forecast")
+Sample Day 1 SPC Forecast: (05/18/2017 1200Z Day 1 Forecast) ![picture alt](https://raw.githubusercontent.com/tmaldrsn/SPC-Conv-Forecast/v0.3.4/img/2017-05-18_1200_UTC_Day_1_convective_outlook.gif "SPC Day 1 05/18/2017 1200Z Forecast")
 
-    
+
 Snippet of text file associated with the forecast:  
-![picture alt](https://raw.githubusercontent.com/tmaldrsn/SPC-Conv-Forecast/v0.3.4/text_forecast.png "Raw Forecast Snippet")    
+![picture alt](https://raw.githubusercontent.com/tmaldrsn/SPC-Conv-Forecast/v0.3.4/img/text_forecast.png "Raw Forecast Snippet")    
 Each 8 digit code is associated with a latitude-longitude coordinate, or an indication of continuation from one point on the border to another point somewhere else on the border, and represents a vertex of the polygon in the figure.
 
 
 Associated plot from plot_categorical.py script:  
-![picture alt](https://raw.githubusercontent.com/tmaldrsn/SPC-Conv-Forecast/v0.3.4/Figure_1.png "Reproduced forecast figure")
+![picture alt](https://raw.githubusercontent.com/tmaldrsn/SPC-Conv-Forecast/v0.3.4/img/Figure_1.png "Reproduced forecast figure")
 
 
 Similar plots for the following events and day outlooks can be created as according to the table below:  
@@ -30,7 +30,7 @@ Similar plots for the following events and day outlooks can be created as accord
     * Added separate plotting for days 4-8 forecast
   * Ability to import text files from previous forecasts
   * Added forecast text file lookup
-  
+
 #### Goals for v0.4:
   * Figure out early morning forecast schedule (0600, 1200 forecast timing)
   * Investigate old SPC forecast format and date cutoff
