+++ img/2017-05-18_1200_UTC_Day_1_convective_outlook.gif
