+++ notnow/checkpolygon.py
