+++ source/get_forecast_object.py
@@ -10,7 +10,7 @@ format = '%(asctime)-15s %(filename)s %(funcName)s %(message)s'
 logging.basicConfig(filename='errlog.log', level=logging.DEBUG, format=format)
 
 
-def get_info(days=1, url=None):
+def main(days=1, url=None):
     """
     Retrieves and formats data from the current SPC forecast online.
     Also able to get information from any archived forecast URL.
@@ -20,6 +20,8 @@ def get_info(days=1, url=None):
         logging.info("No URL provided, extracting most recent available forecast.")
         day = geturl.geturl(days)[0]
         url = geturl.geturl(days)[1]
+
+        print(day)
     else:
         logging.info("URL provided, attempting to extract archived forecast.")
         day = days
@@ -45,7 +47,7 @@ def get_info(days=1, url=None):
         probs = {"tornado":         ["0.02", "0.05", "0.10", "0.15", "0.30", "0.45", "0.60", "SIGN"],
                  "hail":            ["0.05", "0.15", "0.30", "0.45", "0.60", "SIGN"],
                  "wind":            ["0.05", "0.15", "0.30", "0.45", "0.60", "SIGN"],
-                 "categorical":     ["TSTM", "MRGL", "SLGT", "ENH", "MOD", "HIGH"]}
+                 "categorical":     ["TSTM", "MRGL", "SLGT", "ENH", "MDT", "HIGH"]}
         logging.info('Day 1 outlook coordinates stored.')
 
     elif day == 2 or day == 3:
@@ -70,71 +72,14 @@ def get_info(days=1, url=None):
 
         logging.info('Days 4-8 outlook coordinates stored.')
 
-    return coords, probs
+    forecast_object =  {
+        "day": days,
+        "coords": coords,
+        "probs": probs
+    }
+    logging.info("Forecast object created.")
 
+    return forecast_object
 
-def remove_cont(coords_list, substr='99999999'):
-    """
-    Removes '99999999' ==> 'continue' coordinates in the forecast shapes where
-    the outlook shapes reach the border and continue elsewhere on the border.
-    """
-    new_list = []
-    for coord in coords_list:
-        index = 0
-        length = len(substr)
-        while str.find(coord, substr) != -1:
-            index = str.find(coord, substr)
-            new_list.append(coord[0:index])
-            coord = 'TSTM   ' + coord[index+length:]
-        new_list.append(coord[:-1])
-    return new_list
-
-
-def get_coordinates(info, days, event, probability, url=None):
-    """
-    Formats coordinates in form of nested arrays.
-    """
-    
-    coords, probs = info
-    logging.info("Information from retrieval script unpacked.")
-
-    un_coords = remove_cont(coords[event])
-    logging.info("CONTINUE coordinates removed from forecast.")
-
-    d = []
-    for string in un_coords:
-        new_string = list(filter(lambda a: a != '', string.split(' ')))
-        d.append(new_string)
-
-    indices = []
-    for _ in range(len(d)):
-        if len(d[_]) != 0 and d[_][0] in probs[event]:
-            indices.append(_)
-        else:
-            continue
-
-    low = min(i for i in range(len(d)) if len(d[i]) != 0 and d[i][0] == probability)
-    high = max(i for i in range(len(d)) if len(d[i]) != 0 and d[i][0] == probability)
-
-    try:
-        while len(d[high+1]) != 0 and str(d[high+1][0]) not in probs[event]:
-            high += 1
-    except IndexError:
-        pass
-
-    i = low
-    j = -1
-
-    out = []
-    while i <= high:
-        if len(d[i]) != 0 and d[i][0] == probability:
-            j += 1
-            out.append([])
-            new_out = out[j]
-            new_out += d[i][1:]
-        else:
-            new_out += d[i][:]
-        i += 1
-
-    out = [[conversion.convert(coord) for coord in coords] for coords in out]
-    return out
+if __name__=='__main__':
+    main()
